+++ src/main/java/de/th/koeln/archilab/fae/faeteam4service/alarmknopf/Alarmknopfdruck.java
@@ -1,0 +1,13 @@
+package de.th.koeln.archilab.fae.faeteam4service.alarmknopf;
+
+import lombok.Getter;
+
+public class Alarmknopfdruck {
+
+  @Getter
+  private String alarmknopfId;
+
+  public Alarmknopfdruck(String alarmknopfId) {
+    this.alarmknopfId = alarmknopfId;
+  }
+}
