+++ src/main/java/de/th/koeln/archilab/fae/faeteam4service/alarmknopfhilferuf/AlarmknopfHilferuf.java
@@ -1,0 +1,13 @@
+package de.th.koeln.archilab.fae.faeteam4service.alarmknopfhilferuf;
+
+import lombok.Getter;
+
+public class AlarmknopfHilferuf {
+
+  @Getter
+  private String dementiellErkranktePersonId;
+
+  public AlarmknopfHilferuf(String dementiellErkranktePersonId) {
+    this.dementiellErkranktePersonId = dementiellErkranktePersonId;
+  }
+}
