+++ src/main/java/de/th/koeln/archilab/fae/faeteam4service/common/DistanceInMeters.java
@@ -1,0 +1,16 @@
+package de.th.koeln.archilab.fae.faeteam4service.common;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NonNull;
+import lombok.Setter;
+
+@AllArgsConstructor
+public class DistanceInMeters {
+
+  @Getter
+  @Setter
+  @NonNull
+  private double distance;
+
+}
