+++ src/main/java/de/th/koeln/archilab/fae/faeteam4service/common/KafkaConsumer.java
@@ -1,0 +1,16 @@
+package de.th.koeln.archilab.fae.faeteam4service.common;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class KafkaConsumer {
+
+  /**
+  @KafkaListener(topics = "Dummy_Topic",
+      containerFactory = "dementiellErkranktePersonKafkaListenerFactory")
+  public void consumeDementiellErkranktePersonTopic(final String id,
+      final DementiellErkranktePerson dementiellErkranktePerson) {
+
+  }
+  **/
+}
