--- src/main/java/de/th/koeln/archilab/fae/faeteam4service/dto/position/BreitengradDto.java
@@ -1,15 +1,0 @@
-package de.th.koeln.archilab.fae.faeteam4service.dto.position;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-public class BreitengradDto {
-
-  @Getter @Setter @NonNull private double breitengradDezimal;
-
-}
