--- src/main/java/de/th/koeln/archilab/fae/faeteam4service/entities/AlarmknopfHilferuf.java
@@ -1,13 +1,0 @@
-package de.th.koeln.archilab.fae.faeteam4service.entities;
-
-import lombok.Getter;
-
-public class AlarmknopfHilferuf {
-
-  @Getter
-  private String dementiellErkranktePersonId;
-
-  AlarmknopfHilferuf(String dementiellErkranktePersonId) {
-    this.dementiellErkranktePersonId = dementiellErkranktePersonId;
-  }
-}
