--- src/main/java/de/th/koeln/archilab/fae/faeteam4service/entities/DementiellErkranktePersonRepository.java
@@ -1,8 +1,0 @@
-package de.th.koeln.archilab.fae.faeteam4service.entities;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface DementiellErkranktePersonRepository extends
-    CrudRepository<DementiellErkranktePerson, String> {
-
-}
