+++ src/main/java/de/th/koeln/archilab/fae/faeteam4service/position/api/LaengengradDto.java
@@ -1,0 +1,15 @@
+package de.th.koeln.archilab.fae.faeteam4service.position.api;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+public class LaengengradDto {
+
+  @Getter @Setter @NonNull private double laengengradDezimal;
+
+}
