--- src/main/java/de/th/koeln/archilab/fae/faeteam4service/service/KafkaConsumer.java
@@ -1,16 +1,0 @@
-package de.th.koeln.archilab.fae.faeteam4service.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class KafkaConsumer {
-
-  /**
-  @KafkaListener(topics = "Dummy_Topic",
-      containerFactory = "dementiellErkranktePersonKafkaListenerFactory")
-  public void consumeDementiellErkranktePersonTopic(final String id,
-      final DementiellErkranktePerson dementiellErkranktePerson) {
-
-  }
-  **/
-}
