--- src/test/java/de/th/koeln/archilab/fae/faeteam4service/service/AlarmknopfdruckServiceTest.java
@@ -1,48 +1,0 @@
-package de.th.koeln.archilab.fae.faeteam4service.service;
-
-import static org.mockito.Mockito.when;
-
-import de.th.koeln.archilab.fae.faeteam4service.entities.AlarmknopfRepository;
-import de.th.koeln.archilab.fae.faeteam4service.entities.Alarmknopfdruck;
-import de.th.koeln.archilab.fae.faeteam4service.entities.DementiellErkranktePersonRepository;
-import java.util.Optional;
-import org.junit.Test;
-import org.mockito.Mockito;
-import org.springframework.kafka.core.KafkaTemplate;
-
-public class AlarmknopfdruckServiceTest {
-
-  private AlarmknopfdruckService alarmknopfdruckService;
-  private AlarmknopfRepository mockAlarmknopfRepository;
-
-  private DementiellErkranktePersonRepository mockDementiellErkranktePersonRepository;
-  private DementiellErkranktePersonenService dementiellErkranktePersonenService;
-
-  private KafkaTemplate mockKafkaTemplate;
-
-  private static final String ALARMKNOPF_ID = "someId";
-
-  public AlarmknopfdruckServiceTest() {
-    mockAlarmknopfRepository = Mockito.mock(AlarmknopfRepository.class);
-    mockDementiellErkranktePersonRepository = Mockito
-        .mock(DementiellErkranktePersonRepository.class);
-    mockKafkaTemplate = Mockito.mock(KafkaTemplate.class);
-
-    dementiellErkranktePersonenService = new DementiellErkranktePersonenService(
-        mockDementiellErkranktePersonRepository);
-    alarmknopfdruckService = new AlarmknopfdruckService(mockAlarmknopfRepository,
-        dementiellErkranktePersonenService, mockKafkaTemplate);
-  }
-
-  @Test(expected = IllegalArgumentException.class)
-  public void givenHandleKnopfdruckIsCalledAndHilfeknopfIsUnknownIllegalArgumentExceptionShouldBeThrown() {
-
-    Alarmknopfdruck alarmknopfdruck = new Alarmknopfdruck(ALARMKNOPF_ID);
-
-    when(mockAlarmknopfRepository.findById(ALARMKNOPF_ID))
-        .thenReturn(Optional.empty());
-
-    alarmknopfdruckService.handleAlarmknopfdruck(alarmknopfdruck);
-
-  }
-}
